% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukCSSDependency.R
\name{crukCSSDependency}
\alias{crukCSSDependency}
\title{CRUK HTML dependency}
\usage{
crukCSSDependency()
}
\value{
An htmlDependency object
}
\description{
Attaches CSS resources for shinyCRUK components
}
\keyword{internal}
