% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukDatePicker.R
\name{crukDatePicker}
\alias{crukDatePicker}
\title{CRUK Date Picker}
\usage{
crukDatePicker(
  inputId,
  label,
  start,
  end,
  view = "days",
  minView = "days",
  class = "",
  ...
)
}
\arguments{
\item{inputId}{The inputId to access the selected value}

\item{label}{Character. Text to display above the label. Use NULL for no label.}

\item{start}{The minimum date to display. Either a Date object or a string in the form \code{yyyy-mm-dd}.}

\item{end}{The minimum date to display. Either a Date object or a string in the form \code{yyyy-mm-dd}.}

\item{view}{Starting view, one of 'days' (default), 'months' or 'years'.}

\item{minView}{Minimal view, one of 'days' (default), 'months' or 'years'.}

\item{class}{Additional CSS classes to apply to wrapper div.}

\item{...}{Additional arguments passed to \code{shinyWidgets::airDatePIckerInput}}
}
\value{
An HTML div element containing a  styled airDatePickerInput with attached CSS dependencies.
}
\description{
Creates a date picker with CRUK styling. This is a wrapper for \code{shinyWidgets::airDatePickerInput}
with additional CRUK-specific styling applied via CSS dependencies.
}
\examples{
\dontrun{
crukDatePicker(
  inputId = "dateInput",
  label = "Select a date:",
  start = min(data$Date),
  end = max(date$Date),
  view = "months",
  minView = "months"
)

}
}
