% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloWorld.R
\name{HelloWorld}
\alias{HelloWorld}
\title{CRUK Hello World}
\usage{
HelloWorld(exclamation = "!")
}
\arguments{
\item{exclamation}{Additional text to append to "Hello World". Defaults to an exclamation mark}
}
\value{
Some text
}
\description{
Writes Hello World, with optional exclamation mark
}
\examples{
\dontrun{
HelloWorld(

)

}
}
