% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpReview.R
\name{lastReviewHP}
\alias{lastReviewHP}
\title{Health Professional Last Review}
\usage{
lastReviewHP(last_review_date)
}
\arguments{
\item{last_review_date}{A date to display as the last date the page or content was reviewed}
}
\value{
A div containing a styled health professional box and the date of last review
}
\description{
Health Professional Last Review
}
\examples{
hpReview("08 July 2025")

}
