% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpReview.R
\name{lastReviewInternal}
\alias{lastReviewInternal}
\title{Last reviewed date with internal only flag}
\usage{
lastReviewInternal(last_review_date)
}
\arguments{
\item{last_review_date}{A date to display as the last date the page or content was reviewed}
}
\value{
A div containing a styled internal only box and the date of last review to place at the top of pages
}
\description{
Last reviewed date with internal only flag
}
\examples{
lastReviewInternal("08 July 2025")
}
