% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralColumn.R
\name{centralColumn}
\alias{centralColumn}
\title{Central column div}
\usage{
centralColumn(...)
}
\arguments{
\item{...}{}
}
\value{
A div that produces a responsive central column of 900px
}
\description{
Central column div
}
\examples{
centralColumn(
  crukTitle("Page title"),
  p("Some page text")
)
}
