% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukChartTable.R
\name{crukChartTable}
\alias{crukChartTable}
\title{Border-less Card for Chart and Table Combined Output}
\usage{
crukChartTable(chart, table, alt, dataSourceText, dataSourceLink, ...)
}
\arguments{
\item{chart}{A chart output in the form of a \code{shiny::plotOutput},
\code{plotly::plotlyOutput}, or \code{highcharter::highchartOutput}.}

\item{table}{A table output in the form of a \code{shiny::tableOutput},
\code{DT::dataTableOutput}, \code{gt::gt_output}, or \code{reactable::reactableOutput}.}

\item{alt}{Character. Alt text to include alongside the chart. Minimum 15
characters. Please ensure alt text is sufficiently descriptive. Supplied
character string will be appended with the text " Please see table for the data
in an accessible format."}

\item{dataSourceText}{Character or Shiny UI object. Data source text that is displayed at the
bottom of the table. Text is appended with "Data source: " in front.}

\item{dataSourceLink}{Character. URL for the data source hyperlink.}

\item{...}{Optional additional arguments}
}
\value{
A \code{bslib::navset_card_tab} object with attached CSS dependencies
}
\description{
Creates a border-less bslib card with two panels: one for a chart and the other
for a table. This function has a requirement for alt text to ensure accessibility
best practice.
}
