% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukDatePicker.R
\name{crukDatePicker}
\alias{crukDatePicker}
\title{CRUK Date Picker}
\usage{
crukDatePicker(
  inputId,
  label,
  start,
  end,
  view = "months",
  value = Sys.Date(),
  minView = NULL,
  startView = NULL,
  class = "",
  ...
)
}
\arguments{
\item{inputId}{The inputId to access the selected value}

\item{label}{Character. Text to display above the label. Use \code{NULL} for no label.}

\item{start}{The minimum date to display. Either a Date object or a string in the form \code{yyyy-mm-dd}.}

\item{end}{The minimum date to display. Either a Date object or a string in the form \code{yyyy-mm-dd}.}

\item{view}{Starting view, one of 'days', 'months' (default) or 'years'.}

\item{value}{Sets the default selected date. Either a Date object or a string in the form \code{yyyy-mm-dd}. Defaults to \code{Sys.Date()}.}

\item{minView}{Minimal view, one of 'days', 'months' or 'years'. Defaults to equal to \code{view}.}

\item{startView}{View to display when opening the calendar. Defaults to \code{value}}

\item{class}{Additional CSS classes to apply to wrapper div.}

\item{...}{Additional arguments passed to \code{shinyWidgets::airDatePickerInput}}
}
\value{
An HTML div element containing a  styled airDatePickerInput with attached CSS dependencies.
}
\description{
Creates a date picker with CRUK styling. This is a wrapper for \code{shinyWidgets::airDatePickerInput}
with additional CRUK-specific styling applied via CSS dependencies.
}
\examples{
\dontrun{
crukDatePicker(
  inputId = "dateInput",
  label = "Select a date:",
  start = min(data$Date),
  end = max(date$Date),
  view = "months",
  minView = "months"
)

}
}
