% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukFooter.R
\name{crukFooter}
\alias{crukFooter}
\title{CRUK footer}
\usage{
crukFooter()
}
\value{
a div containing a styled footer with links and logos
}
\description{
CRUK footer
}
\examples{
crukFooter()

if (interactive()) {
  crukFooter()
}
}
