% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{crukLogoWide}
\alias{crukLogoWide}
\title{CRUK logo wide}
\usage{
crukLogoWide(height = "50px", width = "auto", class = "")
}
\arguments{
\item{height}{The height of the logo, default is 50px. Must be a valid css unit.}

\item{width}{The width of the logo, may not work unless a height is specified too. Must be a valid css unit.}

\item{class}{Add extra CSS classes, if adding multiple then use c() format}
}
\value{
A wide version of the CRUK as a .png.
}
\description{
This function returns the wide version of the CRUK logo with the css class "cruk-logo-wide". Use sparingly and only when stacked logo doesn't work.
}
\details{
By default the height is set to 50px, but you can specify a width or height.
}
\examples{
crukLogoWide()
}
