% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos.R
\name{crukLogoWide}
\alias{crukLogoWide}
\title{CRUK logo wide}
\usage{
crukLogoWide()
}
\value{
A wide version of the CRUK as a .svg. Use sparingly and only when stacked logo doesn't work
}
\description{
CRUK logo wide
}
\examples{
crukLogoWide()
}
