% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{crukPickerInput}
\alias{crukPickerInput}
\title{CRUK pickerInput}
\usage{
crukPickerInput(
  inputId,
  label,
  choices,
  class = "",
  livesearch = TRUE,
  placeholder = NULL,
  ...,
  options = list()
)
}
\arguments{
\item{options}{}
}
\description{
CRUK pickerInput
}
