% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{crukSelectInput}
\alias{crukSelectInput}
\title{CRUK selectInput}
\usage{
crukSelectInput(inputId, label, choices, selectize = FALSE, class = "", ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{List of values to select from. If elements of the list are named,
then that name rather than the value is displayed to the user.}

\item{selectize}{Whether to use selectize.js or not. Default is FALSE.}

\item{class}{Additional CSS classes to apply to the wrapper div.}

\item{...}{Additional arguments passed to \code{shiny::selectInput}.}
}
\value{
An HTML div element containing a styled select input with attached
  CSS dependencies.
}
\description{
Creates a styled select input control with CRUK branding and custom CSS styling.
This is a wrapper around \code{shiny::selectInput} with additional CRUK-specific
styling applied via CSS dependencies.
}
\examples{
\dontrun{
# In your Shiny UI
crukSelectInput(
  inputId = "biscuit_choice",
  label = "Select your favourite biscuit:",
  choices = c("Hobnob", "Digestive", "Bourbon")
)
}
}
