% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{crukSelectInput}
\alias{crukSelectInput}
\title{CRUK selectInput}
\usage{
crukSelectInput(inputId, label, choices, selectize = FALSE, class = "", ...)
}
\arguments{
\item{...}{}
}
\description{
CRUK selectInput
}
