% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukTheme.R
\name{crukTheme}
\alias{crukTheme}
\title{CRUK Theme}
\usage{
crukTheme(selectors = FALSE)
}
\arguments{
\item{selectors}{TRUE/FALSE whether the navbar includes selectors}
}
\value{
Adds CRUK themeing via CSS
}
\description{
CRUK Theme
}
\examples{
htmltools::tags$head(
  crukTheme()
)
}
