% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukTitle.R
\name{crukTitle}
\alias{crukTitle}
\title{Page Title}
\usage{
crukTitle(Title, Subheading = "")
}
\arguments{
\item{Title}{The title text to display as a h1}

\item{Subheading}{The subheading text to display as a h2}
}
\value{
A div containing a title, subheading and horizontal rule in CRUK styling
}
\description{
Page Title
}
\examples{
crukTitle("Cancer Incidence")

crukTitle(
  "Incidence for common cancers",
  "Latest cancer incidence statistics for the years 2017-2019"
)
}
