% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crukValueBox.R
\name{crukValueBox}
\alias{crukValueBox}
\title{CRUK value box}
\usage{
crukValueBox(
  id,
  title = "",
  value = "",
  icon = "",
  icon_colour = "#00007e",
  detail = ""
)
}
\arguments{
\item{id}{id for the value box}

\item{title}{the title}

\item{value}{the value}

\item{icon}{the icon of the value box}

\item{icon_colour}{the colour of the icon}

\item{detail}{the detail of the value box}
}
\value{
a branded CRUK value box
}
\description{
CRUK value box
}
\examples{
crukValueBox(
  id = "valuBox1",
  title = "Title of the first value box",
  value = "250",
  icon = shiny::icon("weight-scale"),
  detail = "detail text"
)
}
