% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastReview.R
\name{lastReview}
\alias{lastReview}
\title{Last reviewed date with no tag}
\usage{
lastReview(last_review_date)
}
\arguments{
\item{last_review_date}{A date to display as the last date the page or content was reviewed}
}
\value{
A div containing the date of last review to place at the top of pages
}
\description{
Last reviewed date with no tag
}
\examples{
lastReview("08 July 2025")
}
